Move classes to cave package. Did some formatting.
8) Always organize your code into a package hierarchy. With only a few classes, you may not think you will need it, but trust me, it will be very helpful. A typical package name might be something like com.barrybecker4.simulation. Notice that the parts of the name go from most general to specific. The package corresponds to the folder hierarhcy. In your case, I think you should at least create a "cave" package. As you add featrues, you may find that you  need subpackages within the "cave" package like cave.model or cave.ui.
 9) Formatting cleanup within CaveMap code. Whitespace makes the code easier to read. Put a space after the "if" and "for" reserved words, and around binary operators. When you  Whatever formatting style you adopt - be consistent!
 10) Indenting. I prefer to use 4 spaces instead of tabs. I recommend that you turn no "show whitespace" in File | Settings | Appearance and see that you have mixed tabs and spaces everywhere. This looks messy when looking in an editor that might show tabs with a different width. Use either all tabs or all spaces but do not mix them. To fix, I went to File | Settings | Editor | Code Style | Java and checked "use tab character" in the first tab. Then I selected Code | Reformat Code and walah! Intellij made all the code look beautiful!
